fixed comment functionality and deleted unneeded files
